requirements.txt
fastapi
pydantic
uvicorn
bcrypt
python-jose
python-multipart
sqlmodel
mysql-connector-python
cryptography
mariadb
